requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dash==1.7.0
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Compress==1.4.0
flatten-json==0.1.7
future==0.18.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.0
pandas==0.25.3
plaid-python==3.4.0
plotly==4.4.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==1.5
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.2
requests==2.22.0
retrying==1.3.3
scipy==1.4.1
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
